Refactor renderElements function
