+++ src/styles.scss
@@ -22,6 +22,10 @@ $icons: #6b7280;
   display: none;
 }
 
+.crossed {
+  text-decoration: line-through;
+}
+
 /* body */
 
 body {
